Make gtk_container_propagate_draw work with frame window
gtk_container_propagate_draw was not working correctly in this
situation, since it was assuming that the child window is a
descendent of gtk_widget_get_window (container). As a consequence,
we were first adding the toplevel position to the offset, and
then resetting the offset to zero, both of which are not correct.
To fix this, exit the loop when we are hitting a toplevel window.