csd: Fix gtk_window_get_remembered_size to never grow
gtk_window_get_remembered_size returns a size for the content
area, to which we then add the frame size to get the new size
for the frame window. To avoid accidentally growing the window,
return a remembered size that comes from subtracting the frame
size from the frame_window size.