MSVC Builds: Fix gengir Projects
Like the install projects that were fixed few days ago, the gengir projects
did not have info on the intermediate and output directories as a result of
the split up of the property sheets.  Fix this by including the appropriate
property sheet in the gtk-gengir property sheet so that we can avoid
confusing messages from Visual Studio on whether to reload the gengir
project as it was modified, at least on 2008.