Visual C++ Projects: Add rules to generate GTK DBus sources
Since commit c66f194d the gtkdbusgenerated.[c|h] are not included in the
dist tarball and thus have to be generated, which broke the Visual C++
builds.

This patch adds custom build rules for the Visual C++ projects so that
gtkdbusgenerated.[c|h] will be generated upon building the GTK+ DLL
sources, so that the build will be fixed.

This also tells people building GTK+ from the projects that they need
to have Python 2/3 installed and the Python interpretor needs to be in
their PATH before building GTK+ from the projects.