Fix problems with dialog sizing
Many dialogs contain wrapping labels, but don't set
max-width-chars on them. Previously, we were capping their
width at 640, but since 3.10.5, they extend all the way to
the width of the screen, which is not the desired behaviour.

Go back to capping the width of dialogs at 640 in the stable
series. In git master, we will set max-width-chars on the
labels instead.

https://bugzilla.gnome.org/show_bug.cgi?id=719516
https://bugzilla.gnome.org/show_bug.cgi?id=719569