Rework the Visual Studio 2008 Projects a bit
Add a utility project to get config.h and gdkconfig.h from their *.h.win32
(or win32_broadway, if applicable) counterparts, using custom build rules,
so that these "generated" files can also be removed on clean and
"regenerated" upon update.  This also enables the removal of configs in
certain projects that isn't really needed as a result.

Also update and merge the projects and property sheets to include a single
property sheet that it needs, which will then in turn include the other
property sheets that is needed, so that things are cleaner.

Updates to the Visual Studio 2010 projects will follow later.