configure: use CFLAGS, etc. _FOR_BUILD as well
When not doing cross-builds, use the values of CFLAGS, CPPFLAGS and
LDFLAGS as the default value for CFLAGS_FOR_BUILD, CPPFLAGS_FOR_BUILD
and LDFLAGS_FOR_BUILD, respectively.

This avoids having to manually specify these variables in order to get
extract-strings to build properly.

This should really be handled by ax_prog_cc_for_build.m4.  That has been
reported upstream.  This is a workaround for now.

https://bugzilla.gnome.org/show_bug.cgi?id=721346